create rust plumbing for wasm function call
