+++ quadratic-core/src/controller/user_actions/mod.rs
@@ -4,6 +4,7 @@ pub mod auto_complete;
 pub mod borders;
 pub mod cells;
 pub mod clipboard;
+pub mod code;
 pub mod formatting;
 pub mod import;
 pub mod sheets;
